Fix bug which stopped error message being displayed
Fixes bug which meant that if a user cancelled the dialog to find the
configutation file, then the user just saw a blank screen with no
message.

The screen size has been reduced to accommodate smaller monitors. And
the background is completely redrawn at the start so that the text
output doesn't appear in lighter grey boxes - which looked odd.