Improved output, added timestamp and option to overwrite output file
Minor improvements to initial RGB information to make it clearer to
read.

Add timestamp at start of output.

Added option append_to_output_file which if missing/true will add the
output of the current test run to the output file. If set to false, will
overwrite the contents of the output file.